pyproject.toml
[project]
name = "finance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.123.10",
    "google-adk>=1.23.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "uvicorn>=0.40.0",
    "yfinance>=0.2.52",
]
